기능추가, 테스트 : Member R API 개발, MemberTeam 도입에 따른 전체 로직 및 테스트 코드 수정

+++ src/main/java/kr/co/co_working/member/service/MemberService.java
@@ -2,14 +2,19 @@ package kr.co.co_working.member.service;
 
 import ch.qos.logback.core.util.StringUtil;
 import kr.co.co_working.member.dto.MemberRequestDto;
+import kr.co.co_working.member.dto.MemberResponseDto;
+import kr.co.co_working.member.repository.MemberDslRepository;
 import kr.co.co_working.member.repository.MemberRepository;
 import kr.co.co_working.member.repository.entity.Member;
+import kr.co.co_working.memberTeam.service.MemberTeamService;
 import kr.co.co_working.team.repository.TeamRepository;
 import kr.co.co_working.team.repository.entity.Team;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+
+import java.util.List;
 import java.util.NoSuchElementException;
 import java.util.Optional;
 
@@ -17,8 +22,10 @@ import java.util.Optional;
 @Service
 @RequiredArgsConstructor
 public class MemberService {
+    private final MemberTeamService memberTeamService;
     private final MemberRepository repository;
     private final TeamRepository teamRepository;
+    private final MemberDslRepository dslRepository;
 
     /**
      * createMember : Member 등록
@@ -28,30 +35,29 @@ public class MemberService {
      * @throws Exception
      */
     public String createMember(MemberRequestDto.CREATE dto) throws NoSuchElementException, Exception {
-        // 1. ID에 해당하는 Team 조회
-        Optional<Team> selectedTeam = teamRepository.findById(dto.getTeamId());
-
-        // 2. 부재 시 예외 처리
-        if (selectedTeam.isEmpty()) {
-            throw new NoSuchElementException("등록하려는 팀이 존재하지 않습니다. ID : " + dto.getTeamId());
-        }
-        Team team = selectedTeam.get();
-
-        // 3. Member 빌드
+        // 1. Member 빌드
         Member member = Member.builder()
                 .email(dto.getEmail())
                 .password(dto.getPassword())
                 .name(dto.getName())
                 .description(StringUtil.nullStringToEmpty(dto.getDescription()))
-                .team(selectedTeam.get())
                 .build();
 
         // 4. Member 등록
         repository.save(member);
-        team.insertMember(member);
 
         // 5. Email 반환
         return member.getEmail();
+    }
+
+    /**
+     * readMemberList : Member 조회
+     * @param dto
+     * @return
+     * @throws Exception
+     */
+    public List<MemberResponseDto> readMemberList(MemberRequestDto.READ dto) throws Exception {
+        return dslRepository.readMemberList(dto);
     }
 
     /**
@@ -78,9 +84,17 @@ public class MemberService {
             throw new NoSuchElementException("수정하려는 팀이 존재하지 않습니다. ID : " + dto.getTeamId());
         }
 
-        // 5. Member 수정
+        // 5. 객체 추출
         Member member = selectedMember.get();
-        member.updateMember(dto.getName(), dto.getDescription(), selectedTeam.get());
+        Team team = selectedTeam.get();
+
+        // 6. Member 수정
+        member.updateMember(dto.getName(), dto.getDescription());
+
+        // 7. Member 객체에 Team 정보가 있다면 MemberTeam 수정
+        if (member.getMemberTeams() != null) {
+            memberTeamService.updateMemberTeam(member, team);
+        }
     }
 
     @Transactional
@@ -92,18 +106,26 @@ public class MemberService {
         if (selectedMember.isEmpty()) {
             throw new NoSuchElementException("수정하려는 멤버가 존재하지 않습니다. EMAIL : " + dto.getEmail());
         }
+
+        // 3. Member 객체 추출
         Member member = selectedMember.get();
+        
+        // 4. Member 객체에 Team 정보가 있다면
+        if (member.getMemberTeams() != null) {
+            // Team 조회
+            Optional<Team> selectedTeam = teamRepository.findById(dto.getTeamId());
 
-        // 3. Team 조회
-        Optional<Team> selectedTeam = teamRepository.findById(dto.getTeamId());
+            // 부재 시 예외처리
+            if (selectedTeam.isEmpty()) {
+                throw new NoSuchElementException("수정하려는 팀이 존재하지 않습니다. ID : " + dto.getTeamId());
+            }
 
-        // 4. 부재 시 예외처리
-        if (selectedTeam.isEmpty()) {
-            throw new NoSuchElementException("수정하려는 팀이 존재하지 않습니다. ID : " + dto.getTeamId());
+            // Team 객체 추출
+            Team team = selectedTeam.get();
+
+            // Member, MemberTeam 삭제
+            repository.delete(member);
+            memberTeamService.deleteMemberTeam(team);
         }
-        Team team = selectedTeam.get();
-
-        // 5. Member 삭제
-        repository.delete(member);
     }
 }
